pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project]
name = "arepyextras-eo_products"
description = "ArePyExtras EO Products is the Aresys Python module containing utilities to manage external SAR product formats."
authors = [{ name = "Aresys S.R.L.", email = "info@aresys.it" }]
license = { text = "MIT" }
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
]
requires-python = ">= 3.9"
dependencies = [
    "arepytools >= 1.6.1",
    "numpy<2",
    "lxml",
    "tifffile",
    "zarr",
    "h5py",
    "imagecodecs",
    "pyepr"
    ]
dynamic = ["version"]

[project.optional-dependencies]
dev = ['black', 'isort', 'pylint']
doc = [
    "ipython",
    "nbsphinx",
    "sphinx",
    "sphinx-design",
    "sphinx-copybutton",
    "pydata-sphinx-theme",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[tool.setuptools]
license-files = ["LICENSE.txt"]
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["tests", "tests*"]
namespaces = true

[tool.setuptools.dynamic]
version = { attr = "arepyextras.eo_products.__version__" }

[tool.pylint.MASTER]
fail-under = 8.0

[tool.pylint."MESSAGES CONTROL"]
disable = ["import-error", "logging-fstring-interpolation", "logging-not-lazy"]

[tool.pylint.FORMAT]
max-line-length = 120

[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
